refactor and remove resource
