+++ invenio_notify/resources/__init__.py
@@ -1,18 +1,18 @@
+from .basic_db_resource import BasicDbResource
 from .config import (
     NotifyInboxResourceConfig,
     ReviewerResourceConfig,
     InboxApiResourceConfig,
     EndorsementRequestResourceConfig,
 )
-from .resource import (
-    NotifyInboxResource,
-    ReviewerResource,
-    InboxApiResource,
-    EndorsementRequestResource,
-)
+from .endorsement_request_resource import EndorsementRequestResource
+from .inbox_api_resource import InboxApiResource
+from .notify_inbox_resource import NotifyInboxResource
+from .reviewer_resource import ReviewerResource
 
 __all__ = [
     # Resources
+    "BasicDbResource",
     "NotifyInboxResource", 
     "ReviewerResource",
     "InboxApiResource",
