+++ invenio_notify/resources/notify_inbox_resource.py
@@ -1,0 +1,17 @@
+from flask_resources import route
+
+from .basic_db_resource import BasicDbResource
+
+
+class NotifyInboxResource(BasicDbResource):
+
+    def create_url_rules(self):
+        """Create the URL rules for the record resource."""
+        routes = self.config.routes
+        return [
+            # route("POST", routes["list"], self.create),
+            route("GET", routes["item"], self.read),
+            route("GET", routes["list"], self.search),
+            route("DELETE", routes["item"], self.delete),
+            # route("PUT", routes["item"], self.update),
+        ]
